requirements.txt
streamlit>=1.18
transformers>=4.30
langchain>=0.0.190
langchain-community>=0.0.25
sentencepiece>=0.1.99
PyPDF2>=3.0.1
python-docx>=0.8.11
torch>=2.0.1
altair>=4.2.0
vega_datasets>=0.9.0
